Add files via upload
Sunday Times Teaser 2882- From Dec 2017
"Snow white places three balls in a hat. Written on each ball was a different non-zero digit. The balls were removed from the hat in a specific order and used in that order to make a three digit number. This number would be divisible by three but not by seven. She asked the helpers to share out that number of sweets equally amongst the seven of them as far as possible and to give her the remainder. On seeing the remaining umber of sweets she was able to work out the order in which the three digits had been drawn.
What were the three digits?"